requirements.txt
certifi==2021.10.8
chardet==3.0.4
charset-normalizer==2.0.12
click==7.1.2
cmake==3.18.2.post1
cycler==0.11.0
decorator==4.4.2
dlib==19.23.0
docx==0.2.4
docx2txt==0.8
et-xmlfile==1.1.0
exception==0.1.0
face-recognition==1.3.0
face-recognition-models==0.3.0
fonttools==4.29.1
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.2
kiwisolver==1.3.2
lxml==4.8.0
matplotlib==3.5.1
moviepy==1.0.3
numpy==1.22.2
opencv-python==4.5.5.62
openpyxl==3.0.9
packaging==21.3
pandas==1.4.1
Pillow==8.0.1
proglog==0.1.9
pyparsing==3.0.7
python-dateutil==2.8.2
python-docx==0.8.11
python-resize-image==1.1.20
pytz==2021.3
requests==2.24.0
six==1.16.0
tqdm==4.51.0
urllib3==1.25.11
wincertstore==0.2
